# server/python-flask-server-generated-v04/swagger_server/models/inkomen_bri_inkomens2019.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class InkomenBriInkomens2019(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, inkomen: int=None, grondslag: str=None, status: str=None):  # noqa: E501
        """InkomenBriInkomens2019 - a model defined in Swagger

        :param inkomen: The inkomen of this InkomenBriInkomens2019.  # noqa: E501
        :type inkomen: int
        :param grondslag: The grondslag of this InkomenBriInkomens2019.  # noqa: E501
        :type grondslag: str
        :param status: The status of this InkomenBriInkomens2019.  # noqa: E501
        :type status: str
        """
        self.swagger_types = {
            'inkomen': int,
            'grondslag': str,
            'status': str
        }

        self.attribute_map = {
            'inkomen': 'inkomen',
            'grondslag': 'grondslag',
            'status': 'status'
        }
        self._inkomen = inkomen
        self._grondslag = grondslag
        self._status = status

    @classmethod
    def from_dict(cls, dikt) -> 'InkomenBriInkomens2019':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The inkomen_bri_inkomens_2019 of this InkomenBriInkomens2019.  # noqa: E501
        :rtype: InkomenBriInkomens2019
        """
        return util.deserialize_model(dikt, cls)

    @property
    def inkomen(self) -> int:
        """Gets the inkomen of this InkomenBriInkomens2019.


        :return: The inkomen of this InkomenBriInkomens2019.
        :rtype: int
        """
        return self._inkomen

    @inkomen.setter
    def inkomen(self, inkomen: int):
        """Sets the inkomen of this InkomenBriInkomens2019.


        :param inkomen: The inkomen of this InkomenBriInkomens2019.
        :type inkomen: int
        """

        self._inkomen = inkomen

    @property
    def grondslag(self) -> str:
        """Gets the grondslag of this InkomenBriInkomens2019.


        :return: The grondslag of this InkomenBriInkomens2019.
        :rtype: str
        """
        return self._grondslag

    @grondslag.setter
    def grondslag(self, grondslag: str):
        """Sets the grondslag of this InkomenBriInkomens2019.


        :param grondslag: The grondslag of this InkomenBriInkomens2019.
        :type grondslag: str
        """

        self._grondslag = grondslag

    @property
    def status(self) -> str:
        """Gets the status of this InkomenBriInkomens2019.


        :return: The status of this InkomenBriInkomens2019.
        :rtype: str
        """
        return self._status

    @status.setter
    def status(self, status: str):
        """Sets the status of this InkomenBriInkomens2019.


        :param status: The status of this InkomenBriInkomens2019.
        :type status: str
        """

        self._status = status